move: lotto -> lotto.domain 으로 모두 이동
