+++ src/main/java/lotto/domain/LotteryUtil.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.HashSet;
 import java.util.List;
