+++ src/main/java/lotto/ui/InputView.java
@@ -1,0 +1,4 @@
+package lotto.ui;
+
+public class InputView {
+}
