+++ src/test/java/lotto/domain/LotteryRankTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LotteryRank;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
