+++ src/test/java/lotto/domain/LotteryTest.java
@@ -1,5 +1,8 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Lottery;
+import lotto.domain.LotteryAnswer;
+import lotto.domain.LotteryNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
